Move the url type from lingo, lingo#245
