+++ arches_component_lab/src/arches_component_lab/widgets/types.ts
@@ -44,3 +44,8 @@ export interface FileReference {
     description: string;
     title: string;
 }
+
+export interface Url {
+    url: string;
+    url_label: string;
+}
